searching for opponent

+++ v0/sumo_robot/app/inc/ultrasonic_sensor.h
@@ -6,6 +6,6 @@
 #include "int_to_str.h"
 void tim3_init(void);
 void tim3_reset(void);
-float tim3_get_distance(uint16_t tim3_value);
+int tim3_get_distance(uint16_t tim3_value);
 void send_distance_via_uart(uint16_t tim3_value);
 #endif
